+++ modules/zim-cache-redis4cats/src/main/scala/org/bitlap/zim/cache/redis4cats/CatsRedisConfiguration.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.bitlap.zim.cache.catsredis
+package org.bitlap.zim.cache.redis4cats
 import cats.effect.{ IO, Resource }
 import com.typesafe.config.{ Config, ConfigFactory }
 import dev.profunktor.redis4cats._
